Not using xml for its ineded purpose
This file is not reflecting the structure of data but rather reusing html tags as xml elements. This not the purpose of xml.  Maybe the contact part is making use of xml properly to some extent. Find out what data your client like to maintain and display and how this data may be structured. Are there any products that you could list with xml?